move codes to contrib and add tests
